Convert `class Event` to a template.
Two type tags have been created: `vwpp::TaskSignal` and `vwpp::IntSignal`
to indicate the environment in which the event is being used. These tags
are used as the template parameter for `class Event<>` to specialize the
template. `TaskSignal` indicates that a task will wake up any tasks
waiting on the event. `IntSignal` indicates that an interrupt routine
will do the signalling. In the latter case, the `wait()` method takes
an additional `vwpp::IntLock` parameter which guarantees the caller has
properly blocked interrupts.

The Wiki page for this library shows an example why this is the correct
behavior when dealing with interrupt sources.