Add a version `namespace`.
The top-level namespace for this library has always been `vwpp`. This
commit adds a nested namespace, `v2_7`, so that all identifiers'
name-mangling includes this version number. With each release, the
version namespace will change.

By doing this, if someone builds with one header but loads a different
version of the library, they'll get unresolved symbols when loading it
in VxWorks.