Mark `set` methods as `const`.
It may seem wrong to mark a method as `const` when it modifies
hardware. In this case, it's correct; the `const` is referring
to the `VME::Memory` object, not the hardware registers. The
hardware is accesses through a pointer in the object and the
target of the pointer is marked `volatile` so we get the correct
semantics.