Update `VME::Memory` API.
v2.7 introduces a cleaner API for describing hardware registers and
accessing them through the `Memory` class.

Rather than specifying all the register parameters in each call to
`get<>()` and `set<>()`, this API now introduces a register "trait"
template that defines the property of a register. The `get`/`set`
methods use the traits to generate the correct code.