Enable `.set_field()` method for readable registers.
The `.set_field()` method provides an efficient way to write a subset
of bits of a register by reading it, merging with the new value, and
writing it out.  This was even allowed for registers that read
destructively.

This commit adds a template specialization that only allows
non-destructive read registers to support this method. All other
register types will generate a compile-time error.