:recycle: add address space when defining registers
While using this library with an Industry Pack that had A16 and A32
regions, I realized it was possible to specify register access in the
wrong bank of memory.

This commit changes the register template parameters to include the target
address space. Now you get a compile-time error if you try to access the
register in the wrong bank of memory.

Note: if you have two A16 banks, it will allow A16 register definitions to
work in either bank, which is probably incorrect. However, I believe
hardware with two separate banks in the same address space are rare to
non-existent so I think closing this loophole isn't worth the effor.